[Phi] Migrate box coder to phi.
