+++ paddle/phi/api/yaml/legacy_api.yaml
@@ -326,6 +326,16 @@
   kernel :
     func : bitwise_xor
 
+# box_coder
+- api : box_coder
+  args : (Tensor prior_box, Tensor prior_box_var, Tensor target_box, str code_type, bool box_normalized, int axis, float[] variance)
+  output : Tensor(output_box)
+  infer_meta :
+    func : BoxCoderInferMeta
+  kernel :
+    func : box_coder
+  optional : prior_box_var
+
 # brelu
 - api : brelu
   args : (Tensor x, float t_min, float t_max)
